pyproject.toml
[tool.poetry]
name = "sampleapp"
version = "0.1.0"
readme = "README.md"
description = "A sample Python app"
authors = ["Peter Deme"]

[tool.poetry.dependencies]
python = "^3.6"
fastapi = "0.42.0"
uvicorn = "0.10.3"
alembic = "1.3.0"
python-dotenv = "0.10.3"
gunicorn = { version = "20.0.0", optional = true }
gino = "0.8.4"
psycopg2-binary = "2.8.4"

[tool.poetry.extras]
hosting = ["gunicorn"]

[tool.poetry.dev-dependencies]
pytest = "*"
mypy = "*"
mock = "*"
flake8 = "*"
flake8-quotes = "*"
autopep8 = "*"
tox = "*"
rope = "*"
isort = "*"
requests = "*" # Required for Starlette's TestClient
pytest-asyncio = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"